# docs/swagger.yaml
definitions:
  models.CreateUserPayload:
    properties:
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
  models.CreateUserResponse:
    properties:
      id:
        type: integer
      success:
        type: boolean
    type: object
  models.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  models.UserList:
    properties:
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
      password:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
info:
  contact: {}
  description: This is a Messaging app backend.
  title: Your API Title
  version: "1.0"
paths:
  /create-user:
    post:
      consumes:
      - application/json
      description: Accepts a JSON payload to create a new user and returns the new
        user ID
      parameters:
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.CreateUserPayload'
      produces:
      - application/json
      responses:
        "201":
          description: Created user ID
          schema:
            $ref: '#/definitions/models.CreateUserResponse'
      summary: Create a new user
      tags:
      - users
  /users:
    get:
      description: Returns a list of all users (JWT-protected)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.UserList'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get all users
      tags:
      - users
swagger: "2.0"