# internal/models/response.go
package models

import "time"

type CreateUserResponse struct {
	ID      int  `json:"id"`
	Success bool `json:"success"`
}

type UserList struct {
	ID        int       `json:"id"`
	FirstName string    `json:"first_name"`
	LastName  string    `json:"last_name"`
	Email     string    `json:"email"`
	Username  string    `json:"username"`
	Password  string    `json:"password"`
	CreatedAt time.Time `json:"created_at"`
	UpdatedAt time.Time `json:"updated_at"`
}